+++ routes/api.php
@@ -4,6 +4,15 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::group(['namespace' => '\App\Http\Controllers\Api\V1'], function () {
+
+    /*
+     * Chat Server Api
+     */
+    Route::group(['namespace' => 'ChatServerApi'], function () {
+        Route::get('chatServerApi/users/{id}/online', 'UserController@doOnline');
+        Route::get('chatServerApi/users/{id}/offline', 'UserController@doOffline');
+    });
+
     /*
     * General
     */
@@ -53,6 +62,7 @@ Route::group(['namespace' => '\App\Http\Controllers\Api\V1'], function () {
 
             // Contacts
             Route::get('contacts', "ContactController@index");
+            Route::post('contacts/resetNewMessageInfo', "ContactController@resetContactNewMessageInformation");
 
             // Messages
             Route::get('messages', "MessageController@index");