+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+
+## [dev]
+
+### Add
+- add field is_online at users table default value 0
+- add QUEUE_CONNECTION=database to env
+- add jobs table
+- add pending_notifications table
+
 ## [dev,test,master] 
 
 ### CMD
@@ -5,6 +14,7 @@
 - php artisan db:seed BadgesTableSeeder
 - add column badge_id at users table default value 1
 - booking_locations zip and city allow null
+